feat: add skill tree legends component
