+++ src/modules/skill-tree/render-node/render-node.ts
@@ -1,16 +1,4 @@
-import type { CircularSkillNode } from "../skill-tree.type";
-
-export const LEVEL_COLORS: Record<number, { node: string; border: string; connection: string }> = {
-  0: { node: "#312e81", border: "#818cf8", connection: "#4f46e5" }, // Indigo - Centre
-  1: { node: "#065f46", border: "#34d399", connection: "#10b981" }, // Emerald - Niveau 1
-  2: { node: "#1d4ed8", border: "#60a5fa", connection: "#3b82f6" }, // Blue - Niveau 2
-  3: { node: "#9a3412", border: "#fb923c", connection: "#f97316" }, // Orange - Niveau 3
-  4: { node: "#581c87", border: "#c084fc", connection: "#a855f7" }, // Purple - Niveau 4
-  5: { node: "#854d0e", border: "#facc15", connection: "#eab308" }, // Amber - Niveau 5
-  6: { node: "#be123c", border: "#f87171", connection: "#ef4444" }, // Red - Niveau 6
-  7: { node: "#166534", border: "#4ade80", connection: "#22c55e" }, // Green - Niveau 7
-  8: { node: "#7c2d12", border: "#fdba74", connection: "#fb923c" }, // Orange-800 - Niveau 8
-};
+import { LEVEL_COLORS, type CircularSkillNode } from "../skill-tree.type";
 
 const getLevelColors = (level: number) => {
   if (LEVEL_COLORS[level]) {
